stackconfig: Variables and outputs can be sensitive/ephemeral
This should eventually have the same meaning as it does for the modules
language under the ephemeral_values language experiment, but as of this
commit is just accepted during decoding and then ignored. The rest of the
functionality will follow in subsequent commits.

There's no specific experiment guard in this case because the stack
configuration language is already experimental as a whole anyway.